+++ nexus/src/app/saga.rs
@@ -18,7 +18,6 @@ use omicron_common::api::external::Error;
 use omicron_common::api::external::ListResult;
 use omicron_common::api::external::LookupResult;
 use omicron_common::api::external::ResourceType;
-use omicron_common::api::internal;
 use omicron_common::bail_unless;
 use std::sync::Arc;
 use steno::DagBuilder;
@@ -61,7 +60,7 @@ impl super::Nexus {
         &self,
         opctx: &OpContext,
         pagparams: &DataPageParams<'_, Uuid>,
-    ) -> ListResult<internal::Saga> {
+    ) -> ListResult<nexus_types::internal_api::views::Saga> {
         // The endpoint we're serving only supports `ScanById`, which only
         // supports an ascending scan.
         bail_unless!(
@@ -74,7 +73,7 @@ impl super::Nexus {
             .saga_list(marker, pagparams.limit)
             .await
             .into_iter()
-            .map(internal::Saga::from)
+            .map(nexus_types::internal_api::views::Saga::from)
             .map(Ok);
         Ok(futures::stream::iter(saga_list).boxed())
     }
@@ -83,12 +82,12 @@ impl super::Nexus {
         &self,
         opctx: &OpContext,
         id: Uuid,
-    ) -> LookupResult<internal::Saga> {
+    ) -> LookupResult<nexus_types::internal_api::views::Saga> {
         opctx.authorize(authz::Action::Read, &authz::FLEET).await?;
         self.sec_client
             .saga_get(SagaId::from(id))
             .await
-            .map(internal::Saga::from)
+            .map(nexus_types::internal_api::views::Saga::from)
             .map(Ok)
             .map_err(|_: ()| {
                 Error::not_found_by_id(ResourceType::SagaDbg, &id)
