+++ nexus/types/Cargo.toml
@@ -9,6 +9,7 @@ anyhow.workspace = true
 chrono.workspace = true
 base64.workspace = true
 dns-service-client.workspace = true
+futures.workspace = true
 newtype_derive.workspace = true
 openssl.workspace = true
 openssl-sys.workspace = true
@@ -16,6 +17,7 @@ openssl-probe.workspace = true
 schemars = { workspace = true, features = ["chrono", "uuid1"] }
 serde.workspace = true
 serde_json.workspace = true
+steno.workspace = true
 uuid.workspace = true
 
 api_identity.workspace = true
