+++ sled-agent/src/sim/server.rs
@@ -100,7 +100,7 @@ impl Server {
                             .hardware
                             .hardware_threads,
                         usable_physical_ram: NexusTypes::ByteCount::try_from(
-                            config.hardware.physical_ram as i64,
+                            config.hardware.physical_ram,
                         )
                         .unwrap(),
                     },
